Introduce ZIP-extraction model loading for production environments
This version does compile to a mod jar and should run in an actual production fabric environment. However, it probably isn't stable. There are still many things that need to be ironed out with the zip-loading solution, key among them being security and letting runners know this mod is writing to their filesystems. Also, this changes ML input generation so it uses the piece classes instead of the names, as in production the piece classes are obfuscated and their names won't match up.